Streaming sound halting problem SOLVED.
Too bad it's not THE halting problem ;-)

Basically, since the OpenAL's renderer confuses a buffer underrun with an end-of-stream condition,
I reworked the OpenAL renderer to keep manual track of the desired playing state of a renderable source.
Which is great: a renderer-specific bug fixed with (mostly) renderer-specific code.

Only change to the scene manager: it performs an idle update of the renderable prior to deactivating
it. Functions as a hook for the renderer to do internal checks, and, perhaps (as in this case), reactivate
the renderable source.